requirements.txt
pillow==10.4.0
tkinterdnd2~=0.4.2